Following requests addressed:
1. Add visual indicator (such as hourglass) to indicate when vparse processing is occurring.
2. Place cursor in Title field by default when program is opened for ease of keyboard entry.
3. Fix Volume up increment skip from “blank/null” to “2”.   Fixed April 10, 2019
4. Add Exit button to File Operations Status pop-up window to exit the vparse program. (OK button will continue to take you back to the main program window to vparse another file). Note: will need further refinement/thought because the plan is to add parsing back to the vparse program. Ken also suggested a user specified default auto exit after successful parse.
5. Clearer language for File Operations Status pop-up window, suggest:

change “No Ext File:” to “Volume file updated:”
change “TXT File:” to “Volume file to HOLDING”
change “DOC File:” to “Volume files archived” (This would mean both the .DOC and the .0yy file. Is it possible to lump those as Volume files archived? We might need a fourth output that says TXT file archived, and call this one DOC file archived.)
Result is either OK (Pass) or Fail.
Note: In the above status result, a new updated volume file was created on L:\??, the updated ttvvmmdd.019 file was sent to HOLDING, but the ttvvmmdd.doc file and ttvvmmdd.019 file was not archived to L:\x??.
6.  Add option in settings to select “DOC” or “DOCX” to support future use of new Word extension format.